Donations Block: do not use admin color on the frontend (#30864)
Fixes #30811

Follow-up from #16687

The Donations block uses a common stylesheet that is used in the block editor and on the frontend, where css vars like `--wp-admin-theme-color` are not available. Let's switch to using a color that will be available on the frontend.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5062805803