Newsletter categories: exposes value of wpcom_newsletter_categories_location to JavaScript (#33374)
* This update exposes the Newsletter category settings load to JavaScript. The settings allow the system to determine the location of the newsletter categories, which can now be modified using the 'wpcom_newsletter_categories_location' filter.

* Update comments for load_newsletter_categories_settings function

Changed the purpose of the function load_newsletter_categories_settings in the comments. It now indicates that the function is to store the information about where the newsletter categories should be shown, not the map provider information as previously stated. The code change reflects this new focus.

* Update comments for load_newsletter_categories_settings function

Changed the purpose of the function load_newsletter_categories_settings in the comments. It now indicates that the function is to store the information about where the newsletter categories should be shown, not the map provider information as previously stated. The code change reflects this new focus.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/6419308100